+++ src/components/ModelizerTextView.vue
@@ -21,8 +21,8 @@
     >
       <template v-slot="{ file }">
         <monaco-editor
-          :content="file.content"
-          :viewType="viewType"
+          :fileInput="file"
+          :project-name="projectName"
         />
       </template>
     </file-tabs>
@@ -41,15 +41,14 @@ import {
   ref,
   watch,
 } from 'vue';
-import { getProjectFiles, readProjectFile } from 'src/composables/Project';
+import { getProjectFiles, readProjectFile, writeProjectFile } from 'src/composables/Project';
 import FileEvent from 'src/composables/events/FileEvent';
 import GitEvent from 'src/composables/events/GitEvent';
+import PluginEvent from 'src/composables/events/PluginEvent';
+import { getPlugins } from 'src/composables/PluginManager';
+import { FileInformation } from 'leto-modelizer-plugin-core';
 
 const props = defineProps({
-  viewType: {
-    type: String,
-    default: 'model',
-  },
   projectName: {
     type: String,
     required: true,
@@ -67,6 +66,7 @@ let createFileSubscription;
 let deleteFileSubscription;
 let updateRemoteSubscription;
 let checkoutSubscription;
+let pluginRenderSubscription;
 
 /**
  * Update fileTabArray array when a new file is open.
@@ -152,26 +152,50 @@ function updateSelectedNode(node) {
  * @param {String} fileName - Name of the file to create.
  */
 function onCreateFileEvent({ name, isFolder }) {
-  return updateProjectFiles().then(() => {
-    if (!isFolder) {
-      const parentNode = selectedNode.value.id === props.projectName ? '' : `${selectedNode.value.id}/`;
-      const newFileId = `${parentNode}${name}`;
-      activeFileTab.value = { isSelected: false, id: newFileId };
+  return updateProjectFiles()
+    .then(() => {
+      if (isFolder) {
+        return Promise.resolve();
+      }
+      return readProjectFile(props.projectName, new FileInformation({ path: name }));
+    })
+    .then((fileInput) => {
+      if (!isFolder) {
+        activeFileTab.value = { isSelected: true, id: fileInput.path };
 
-      FileEvent.OpenFileEvent.next({
-        id: newFileId,
-        label: name.substring(name.lastIndexOf('/') + 1),
-        content: ' ',
+        FileEvent.OpenFileEvent.next({
+          id: fileInput.path,
+          label: name.substring(name.lastIndexOf('/') + 1),
+          content: fileInput.content,
+        });
+      }
+
+      let folder = `${selectedNode.value.id || props.projectName}/${name}`;
+
+      if (folder.indexOf('/') > 0) {
+        folder = folder.substring(0, folder.lastIndexOf('/'));
+      }
+
+      FileEvent.ExpandFolderEvent.next(folder);
+    });
+}
+
+/**
+ * Render components and update files accordingly.
+ */
+function renderPlugins() {
+  const plugins = getPlugins();
+  plugins.forEach((plugin) => {
+    const render = plugin.renderer.render(plugin.components, [], 'new_file.tf');
+
+    render.forEach((file) => {
+      writeProjectFile(props.projectName, file).then(() => {
+        FileEvent.CreateFileEvent.next({
+          name: file.path.substring(file.path.lastIndexOf('/') + 1),
+          isFolder: false,
+        });
       });
-    }
-
-    let folder = `${selectedNode.value.id}/${name}`;
-
-    if (folder.indexOf('/') > 0) {
-      folder = folder.substring(0, folder.lastIndexOf('/'));
-    }
-
-    FileEvent.ExpandFolderEvent.next(folder);
+    });
   });
 }
 
@@ -187,6 +211,7 @@ onMounted(() => {
   deleteFileSubscription = FileEvent.DeleteFileEvent.subscribe(updateProjectFiles);
   updateRemoteSubscription = GitEvent.UpdateRemoteEvent.subscribe(updateProjectFiles);
   checkoutSubscription = GitEvent.CheckoutEvent.subscribe(updateProjectFiles);
+  pluginRenderSubscription = PluginEvent.RenderEvent.subscribe(renderPlugins);
 });
 
 onUnmounted(() => {
@@ -196,6 +221,7 @@ onUnmounted(() => {
   deleteFileSubscription.unsubscribe();
   updateRemoteSubscription.unsubscribe();
   checkoutSubscription.unsubscribe();
+  pluginRenderSubscription.unsubscribe();
 });
 </script>