+++ changelog.md
@@ -36,5 +36,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update PluginEvent for action menu
 - Add fs writeProjectFile & createProjectFolder in Project composable
 - Add create and delete file or folder action
+- Add ParseEvent & RenderEvent management
 
 [unreleased]: https://github.com/ditrit/leto-modelizer/blob/main/changelog.md#unreleased